+++ commons/src/commonMain/kotlin/com/vitorpamplona/amethyst/commons/nip23LongContent/ui/article/ArticleHeader.kt
@@ -18,7 +18,7 @@
  * AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.vitorpamplona.amethyst.commons.ui.article
+package com.vitorpamplona.amethyst.commons.nip23LongContent.ui.article
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
