+++ desktopApp/src/jvmMain/kotlin/com/vitorpamplona/amethyst/desktop/ui/ArticleEditorScreen.kt
@@ -65,9 +65,9 @@ import androidx.compose.ui.unit.sp
 import com.vitorpamplona.amethyst.commons.icons.symbols.Icon
 import com.vitorpamplona.amethyst.commons.icons.symbols.MaterialSymbols
 import com.vitorpamplona.amethyst.commons.model.nip23LongContent.LongFormPublishAction
-import com.vitorpamplona.amethyst.commons.ui.editor.MarkdownEditorState
-import com.vitorpamplona.amethyst.commons.ui.editor.MarkdownToolbar
-import com.vitorpamplona.amethyst.commons.ui.editor.MetadataPanel
+import com.vitorpamplona.amethyst.commons.nip23LongContent.ui.editor.MarkdownEditorState
+import com.vitorpamplona.amethyst.commons.nip23LongContent.ui.editor.MarkdownToolbar
+import com.vitorpamplona.amethyst.commons.nip23LongContent.ui.editor.MetadataPanel
 import com.vitorpamplona.amethyst.commons.ui.markdown.RenderMarkdown
 import com.vitorpamplona.amethyst.desktop.account.AccountState
 import com.vitorpamplona.amethyst.desktop.network.DesktopRelayConnectionManager
